docs-viewer/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
google-cloud-storage>=2.16.0

